# docs/swagger.yaml
basePath: /
definitions:
  model.StatusType:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - StatusTodo
    - StatusInProgress
    - StatusDone
  model.Task:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      status:
        $ref: '#/definitions/model.StatusType'
      title:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для управления задачами
  title: Task Manager API
  version: "1.0"
paths:
  /api/item:
    post:
      consumes:
      - application/json
      description: Добавляет новую задачу
      parameters:
      - description: Задача
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/model.Task'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать задачу
      tags:
      - tasks
  /api/item/{id}:
    delete:
      description: Удаляет задачу по ID
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить задачу
      tags:
      - tasks
    get:
      description: Возвращает задачу по идентификатору
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить задачу по ID
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Обновляет данные задачи
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные задачи
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/model.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить задачу
      tags:
      - tasks
  /api/items:
    get:
      description: Возвращает список всех задач
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Task'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Получить все задачи
      tags:
      - tasks
schemes:
- http
swagger: "2.0"